chore(eslint): migrate to flat config, add next/core-web-vitals, improve ignore patterns
