+++ eslint.config.ts
@@ -1,0 +1,36 @@
+import js from '@eslint/js';
+import globals from 'globals';
+import tseslint from 'typescript-eslint';
+import react from 'eslint-plugin-react';
+import eslintPluginPrettierRecommended from 'eslint-plugin-prettier/recommended';
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
+import { FlatCompat } from '@eslint/eslintrc';
+import eslint from '@eslint/js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+});
+
+export default tseslint.config([
+  { ignores: ['node_modules/', 'dist/', '*.log', '.env', '.next/', 'out/'] },
+  {
+    files: ['**/*.{js,mjs,cjs,ts,jsx,tsx}'],
+    plugins: { js, react },
+    settings: {
+      react: {
+        version: 'detect',
+      },
+    },
+  },
+  { files: ['**/*.{js,mjs,cjs,ts,jsx,tsx}'], languageOptions: { globals: globals.browser } },
+  eslint.configs.recommended,
+  ...tseslint.configs.recommended,
+  ...compat.extends('next/core-web-vitals', 'next/typescript'),
+  react.configs.flat.recommended,
+  react.configs.flat['jsx-runtime'],
+  eslintPluginPrettierRecommended,
+]);
