+++ package.json
@@ -6,7 +6,7 @@
     "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint --config eslint.config.ts",
     "prepare": "husky"
   },
   "dependencies": {
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@commitlint/cli": "^19.8.1",
     "@commitlint/config-conventional": "^19.8.1",
+    "@eslint/eslintrc": "^3.3.1",
     "@eslint/js": "^9.26.0",
     "@tailwindcss/postcss": "^4.1.7",
     "@types/node": "^22.15.18",
